+++ Code/Marching Cubes/src/Geometry/MarchingChunk.cpp
@@ -1,6 +1,6 @@
 #include "MarchingChunk.h"
 #include <iostream>
-MarchingChunk::MarchingChunk(glm::vec3 chunkLocation, glm::vec3 chunkSize, glm::vec3 chunkStride, GeometryGenerator* Generator)
+MarchingChunk::MarchingChunk(glm::vec3 chunkLocation, glm::uvec3 chunkSize, glm::vec3 chunkStride, GeometryGenerator* Generator)
    :myLocation(chunkLocation),mySize(chunkSize),myStride(chunkStride),Generator(Generator)
 {
     //ctor
